@@ -1,8 +1,8 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.4.32'
-    ext.dagger_version = "2.13"
+    ext.kotlin_version = '1.5.31'
+    ext.dagger_version = "2.39.1"
     repositories {
         maven {
             url "https://repo.eclipse.org/content/repositories/paho-snapshots/"
@@ -13,13 +13,13 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.2.1'
+        classpath 'com.android.tools.build:gradle:4.2.2'
         classpath "io.realm:realm-gradle-plugin:5.15.1"
-        classpath 'com.google.gms:google-services:4.3.8'
+        classpath 'com.google.gms:google-services:4.3.10'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         // Add the dependency for the Performance Monitoring plugin
-        classpath 'com.google.firebase:perf-plugin:1.3.5'  // Performance Monitoring plugin
-        classpath 'com.google.firebase:firebase-crashlytics-gradle:2.5.2'
+        classpath 'com.google.firebase:perf-plugin:1.4.0'  // Performance Monitoring plugin
+        classpath 'com.google.firebase:firebase-crashlytics-gradle:2.7.1'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
