@@ -18,7 +18,7 @@ import androidx.annotation.Nullable;
 import androidx.core.content.ContextCompat;
 import androidx.fragment.app.FragmentManager;
 import androidx.lifecycle.Observer;
-import androidx.lifecycle.ViewModelProviders;
+import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.DefaultItemAnimator;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
@@ -135,7 +135,7 @@ public class MemberShipPlanActivity extends BaseBindingActivity<MembershipPlanBi
 
     private void callBinding() {
 
-        viewModel = ViewModelProviders.of(this).get(PurchaseViewModel.class);
+        viewModel = new ViewModelProvider(this).get(PurchaseViewModel.class);
         modelCall();
     }
 
