@@ -28,7 +28,7 @@ import androidx.core.widget.ImageViewCompat;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 import androidx.lifecycle.Observer;
-import androidx.lifecycle.ViewModelProviders;
+import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import androidx.viewpager.widget.ViewPager;
@@ -257,7 +257,7 @@ public class TutorialActivity extends BaseBindingActivity<ActivitySeriesDetailBi
 
 
     private void modelCall() {
-        railInjectionHelper = ViewModelProviders.of(this).get(RailInjectionHelper.class);
+        railInjectionHelper = new ViewModelProvider(this).get(RailInjectionHelper.class);
         assetListContinue = new ArrayList<>();
         getBinding().noConnectionLayout.setVisibility(View.GONE);
         getBinding().player.setVisibility(View.VISIBLE);
@@ -270,7 +270,7 @@ public class TutorialActivity extends BaseBindingActivity<ActivitySeriesDetailBi
         preference = KsPreferenceKeys.getInstance();
         isLogin = preference.getAppPrefLoginStatus();
         token = preference.getAppPrefAccessToken();
-        viewModel = ViewModelProviders.of(this).get(SeriesViewModel.class);
+        viewModel = new ViewModelProvider(this).get(SeriesViewModel.class);
         if (isLogin) {
             AppUserModel signInResponseModel = AppUserModel.getInstance();
             if (signInResponseModel != null) {
@@ -353,7 +353,7 @@ public class TutorialActivity extends BaseBindingActivity<ActivitySeriesDetailBi
     private void getSeriesDetail() {
         modelCall();
         postCommentClick();
-        RailInjectionHelper railInjectionHelper = ViewModelProviders.of(this).get(RailInjectionHelper.class);
+        RailInjectionHelper railInjectionHelper = new ViewModelProvider(this).get(RailInjectionHelper.class);
         railInjectionHelper.getSeriesDetailsV2(String.valueOf(seriesId)).observe(TutorialActivity.this, new Observer<ResponseModel>() {
             @Override
             public void onChanged(ResponseModel response) {
@@ -682,7 +682,7 @@ public class TutorialActivity extends BaseBindingActivity<ActivitySeriesDetailBi
 
         }
 
-        getBinding().tabLayout.addOnTabSelectedListener(new TabLayout.BaseOnTabSelectedListener() {
+        getBinding().tabLayout.addOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
             @Override
             public void onTabSelected(TabLayout.Tab tab) {
                 showLoading(getBinding().progressBar, true);
